requirements-new.txt
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1598353661586/work
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.17.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kiwisolver==1.2.0
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mysqlclient==2.0.1
nb-conda-kernels @ file:///C:/ci/nb_conda_kernels_1600427279401/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.5
notebook @ file:///C:/ci/notebook_1601501643625/work
numpy==1.19.3
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
regex==2020.10.28
requests==2.24.0
requests-file==1.5.1
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.3
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.0.2
tornado==6.0.4
tqdm==4.51.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
urllib3==1.25.11
validators==0.18.1
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
WTForms==2.3.3
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
